Add files via upload
In this discord.py bot development tutorial, you will learn how to utilize the Cogs feature of the discord.py library. By the end of this tutorial, you will have an understanding of how to use Cogs to better organize and lay out your code to prevent confusion for yourself or other readers.